requirements.txt
flake8==6.0.0
pytest==7.4.0
selenium==4.10.0
webdriver-manager==3.9.1
pydantic~=1.10.9
delayed-assert~=0.3.6
pycairo~=1.24.0
pygame~=2.5.0
